Added Deck class, modified Card
Card now calculates it's suit and face based on a unique id given by the
deck class.
Considering adding ability to get info as string or as an int depending
on whether it needs printed or to do logic
Added deck class to handle creation and tracking cards.
Deck now handles shuffling cards, can be used to draw and discard.
Drawn from position 0, discarded to highest position.
Deck is static and is initialized upon any function call.
id for card follows the pattern club, diamond, heart, spade, club,
diamond.... and so on. starts at aces and goes up to kings. Ex) id 1 is
ace of clubs, id 10 is 3 of diamonds.